pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "barcode_indexing_tool"
version = "0.0.1"
authors = [
  { name="Tiziano Bacocco", email="tizbac2@gmail.com" },
]
description = "Simple script to index folders of images containing barcodes"
requires-python = ">=3.7"
dependencies=["pyzbar", "Pillow"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
[project.scripts]
barcode_indexing_tool = "barcode_indexing_tool.index_images:main"

[project.urls]
"Homepage" = "https://github.com/pypa/sampleproject"
"Bug Tracker" = "https://github.com/pypa/sampleproject/issues"